moved common layout code to absrtact class
