+++ src/main/java/com/github/onsdigital/logging/layouts/PrettyJsonLayout.java
@@ -2,7 +2,6 @@ package com.github.onsdigital.logging.layouts;
 
 import ch.qos.logback.classic.spi.ILoggingEvent;
 import com.github.onsdigital.logging.layouts.model.JsonLogItem;
-import com.github.onsdigital.logging.layouts.style.Styler;
 
 import java.io.IOException;
 
@@ -14,7 +13,7 @@ public class PrettyJsonLayout extends JsonLayout {
     @Override
     public String doLayout(ILoggingEvent event) {
         try {
-            return Styler.toPrettyJson(new JsonLogItem(event));
+            return toPrettyJson(new JsonLogItem(event));
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
